Some C-API implementation cleanup
* Moved internal class definitions from c4Query.cc into header files
* Moved LOCK() macro into c4Internal.hh
* Moved c4PredictiveQuery stubs to c4PredictiveQuery.cc
